requirements.txt
torch>=1.7.0
torchvision>=0.8.1
opencv-python>=4.1.2
PyYAML>=5.3.1
tqdm>=4.41.0
seaborn>=0.11.0
gradio==3.28.1
Pillow~=9.5.0
numpy~=1.24.3